requirements.txt
ctransformers==0.2.5
sentence-transformers==3.2.1
pinecone-client==5.0.1
langchain==0.0.225
flask
pypdf
python-dotenv
langchain-community==0.3.3
pinecone
setuptools
langchain-huggingface==0.1.0
#sentence-transformers-3.2.1
-e .   #execute setup.py in backend 
